Silenced warnings about logging NSUIntegers by casting and logging %lu rather than %i

+++ wineguide/controllers/RestaurantDetailsVC.h
@@ -11,7 +11,7 @@
 
 @protocol RestaurantDetailsVC_WineSelectionDelegate
 
-- (void)loadWineList:(int)listNumber;
+- (void)loadWineList:(NSUInteger)listNumber;
 
 @end
